# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BarOut": {
        "properties": {
          "close": {
            "title": "Close",
            "type": "number"
          },
          "high": {
            "title": "High",
            "type": "number"
          },
          "is_outlier": {
            "title": "Is Outlier",
            "type": "boolean"
          },
          "is_quote_only": {
            "title": "Is Quote Only",
            "type": "boolean"
          },
          "low": {
            "title": "Low",
            "type": "number"
          },
          "open": {
            "title": "Open",
            "type": "number"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "volume": {
            "title": "Volume",
            "type": "number"
          }
        },
        "required": [
          "ts",
          "open",
          "high",
          "low",
          "close",
          "volume",
          "is_quote_only",
          "is_outlier"
        ],
        "title": "BarOut",
        "type": "object"
      },
      "BarsOut": {
        "properties": {
          "bars": {
            "items": {
              "$ref": "#/components/schemas/BarOut"
            },
            "title": "Bars",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "timeframe",
          "source",
          "count",
          "bars"
        ],
        "title": "BarsOut",
        "type": "object"
      },
      "CalibrationOut": {
        "properties": {
          "computed_at": {
            "format": "date-time",
            "title": "Computed At",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/CalibrationRowOut"
            },
            "title": "Rows",
            "type": "array"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "model_version",
          "computed_at",
          "rows"
        ],
        "title": "CalibrationOut",
        "type": "object"
      },
      "CalibrationRowOut": {
        "properties": {
          "empirical": {
            "title": "Empirical",
            "type": "number"
          },
          "empirical_conformal": {
            "title": "Empirical Conformal",
            "type": "number"
          },
          "first_origin": {
            "format": "date",
            "title": "First Origin",
            "type": "string"
          },
          "horizon_days": {
            "title": "Horizon Days",
            "type": "integer"
          },
          "last_origin": {
            "format": "date",
            "title": "Last Origin",
            "type": "string"
          },
          "n": {
            "title": "N",
            "type": "integer"
          },
          "nominal": {
            "title": "Nominal",
            "type": "number"
          },
          "pinball_baseline": {
            "title": "Pinball Baseline",
            "type": "number"
          },
          "pinball_model": {
            "title": "Pinball Model",
            "type": "number"
          },
          "steps": {
            "title": "Steps",
            "type": "integer"
          }
        },
        "required": [
          "horizon_days",
          "steps",
          "nominal",
          "empirical",
          "empirical_conformal",
          "n",
          "pinball_model",
          "pinball_baseline",
          "first_origin",
          "last_origin"
        ],
        "title": "CalibrationRowOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LevelIn": {
        "properties": {
          "horizon_days": {
            "title": "Horizon Days",
            "type": "integer"
          },
          "level": {
            "exclusiveMinimum": 0.0,
            "title": "Level",
            "type": "number"
          }
        },
        "required": [
          "level",
          "horizon_days"
        ],
        "title": "LevelIn",
        "type": "object"
      },
      "LevelOut": {
        "description": "Two different questions about one price level, kept apart.",
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "ends_above": {
            "title": "Ends Above",
            "type": "number"
          },
          "ends_below": {
            "title": "Ends Below",
            "type": "number"
          },
          "horizon_days": {
            "title": "Horizon Days",
            "type": "integer"
          },
          "level": {
            "title": "Level",
            "type": "number"
          },
          "n_paths": {
            "title": "N Paths",
            "type": "integer"
          },
          "start_price": {
            "title": "Start Price",
            "type": "number"
          },
          "steps": {
            "title": "Steps",
            "type": "integer"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "touches": {
            "title": "Touches",
            "type": "number"
          }
        },
        "required": [
          "symbol",
          "as_of",
          "start_price",
          "level",
          "horizon_days",
          "steps",
          "n_paths",
          "ends_above",
          "ends_below",
          "touches"
        ],
        "title": "LevelOut",
        "type": "object"
      },
...
      "RegimeEvaluationOut": {
        "description": "Walk-forward results: every figure is from days the model had not seen.",
        "properties": {
          "average_run_length": {
            "title": "Average Run Length",
            "type": "number"
          },
          "baseline_log_density": {
            "title": "Baseline Log Density",
            "type": "number"
          },
          "first_test_day": {
            "format": "date",
            "title": "First Test Day",
            "type": "string"
          },
          "last_test_day": {
            "format": "date",
            "title": "Last Test Day",
            "type": "string"
          },
          "model_log_density": {
            "title": "Model Log Density",
            "type": "number"
          },
          "n_days": {
            "title": "N Days",
            "type": "integer"
          },
          "next_day_volatility": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Next Day Volatility",
            "type": "object"
          },
          "volatility_is_ordered": {
            "title": "Volatility Is Ordered",
            "type": "boolean"
          }
        },
        "required": [
          "n_days",
          "first_test_day",
          "last_test_day",
          "next_day_volatility",
          "volatility_is_ordered",
          "model_log_density",
          "baseline_log_density",
          "average_run_length"
        ],
        "title": "RegimeEvaluationOut",
        "type": "object"
      },
      "RegimeModelOut": {
        "properties": {
          "bic_by_states": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Bic By States",
            "type": "object"
          },
          "n_train": {
            "title": "N Train",
            "type": "integer"
          },
          "train_end": {
            "format": "date",
            "title": "Train End",
            "type": "string"
          },
          "train_start": {
            "format": "date",
            "title": "Train Start",
            "type": "string"
          },
          "trained_at": {
            "format": "date-time",
            "title": "Trained At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "trained_at",
          "train_start",
          "train_end",
          "n_train",
          "bic_by_states"
        ],
        "title": "RegimeModelOut",
        "type": "object"
      },
      "RegimeOut": {
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "days_in_state": {
            "title": "Days In State",
            "type": "integer"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RegimeEvaluationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/RegimePoint"
            },
            "title": "History",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "model": {
            "$ref": "#/components/schemas/RegimeModelOut"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Probabilities",
            "type": "object"
          },
          "probability": {
            "title": "Probability",
            "type": "number"
          },
          "states": {
            "items": {
              "$ref": "#/components/schemas/RegimeStateOut"
            },
            "title": "States",
            "type": "array"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "as_of",
          "label",
          "probability",
          "probabilities",
          "days_in_state",
          "states",
          "history",
          "model",
          "evaluation"
        ],
        "title": "RegimeOut",
        "type": "object"
      },
      "RegimePoint": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "probability": {
            "title": "Probability",
            "type": "number"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "ts",
          "label",
          "probability"
        ],
        "title": "RegimePoint",
        "type": "object"
      },
...
      "SimulationOut": {
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "fan": {
            "additionalProperties": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "title": "Fan",
            "type": "object"
          },
          "horizons": {
            "items": {
              "$ref": "#/components/schemas/OutlookHorizon"
            },
            "title": "Horizons",
            "type": "array"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "n_paths": {
            "title": "N Paths",
            "type": "integer"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "start_price": {
            "title": "Start Price",
            "type": "number"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "as_of",
          "start_price",
          "n_paths",
          "seed",
          "model_version",
          "horizons",
          "fan"
        ],
        "title": "SimulationOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Market outlook analytics for Bitcoin, gold, and US stocks. Analytics only: this API never places trades. All timestamps are UTC.",
    "title": "RADAR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/assets/{symbol}/bars": {
      "get": {
        "description": "Stored bars in time order. Without `start`, the latest `limit` bars.",
        "operationId": "get_bars_api_v1_assets__symbol__bars_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "timeframe",
            "required": false,
            "schema": {
              "default": "1Hour",
              "pattern": "^(1Hour|1Day)$",
              "title": "Timeframe",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bars"
      }
    },
    "/api/v1/assets/{symbol}/calibration": {
      "get": {
        "description": "How often the simulator's past ranges contained what happened.",
        "operationId": "get_calibration_api_v1_assets__symbol__calibration_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Calibration"
      }
    },
    "/api/v1/assets/{symbol}/regime": {
      "get": {
        "description": "The current market regime, its history, and how the model has measured.",
        "operationId": "get_regime_api_v1_assets__symbol__regime_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 365,
              "maximum": 5000,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegimeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Regime"
      }
    },
    "/api/v1/assets/{symbol}/simulation": {
      "get": {
        "description": "The latest stored simulator run: the outcome distribution at each horizon.",
        "operationId": "get_simulation_api_v1_assets__symbol__simulation_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Simulation"
      }
    },
    "/api/v1/assets/{symbol}/simulation/level": {
      "post": {
        "description": "Chances of ending beyond, and of touching, a price level. Counted from stored paths.",
        "operationId": "post_simulation_level_api_v1_assets__symbol__simulation_level_post",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LevelIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LevelOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Simulation Level"
      }
    },
...
  }
}